feat: Completed Club Form
This commit marks the completion of the Club Form feature. I have ensured that all required fields are present and functional, and that the form is now able to submit data to the server.

With the completion of this feature, users of the application will be able to create new club entries more easily, and the data collected will be more organized and complete.

The changes made in this commit should improve the overall user experience of the application and contribute to the success of the project.