86 frontendt redone final version 1 push (#87)
* feat: Started Club Project Form

This commit marks the beginning of work on the Club Project Form. I have created the initial files and set up the necessary components to start building the form.

The Club Project Form is an important feature for our application and will allow users to create new projects associated with their clubs. With this commit, I have laid the foundation for this feature and am ready to continue development.

* feat: Started Club Project Form

This commit marks the beginning of work on the Club Project Form. I have created the initial files and set up the necessary components to start building the form.

The Club Project Form is an important feature for our application and will allow users to create new projects associated with their clubs. With this commit, I have laid the foundation for this feature and am ready to continue development.