merge (#89)
* 86 frontendt redone final version 1 push (#88)

* feat: Started Club Project Form

This commit marks the beginning of work on the Club Project Form. I have created the initial files and set up the necessary components to start building the form.

The Club Project Form is an important feature for our application and will allow users to create new projects associated with their clubs. With this commit, I have laid the foundation for this feature and am ready to continue development.

* feat: Started Club Project Form

This commit marks the beginning of work on the Club Project Form. I have created the initial files and set up the necessary components to start building the form.

The Club Project Form is an important feature for our application and will allow users to create new projects associated with their clubs. With this commit, I have laid the foundation for this feature and am ready to continue development.

* feat: Add ClubProject and DSG Project forms and refactor controllers and frontend components

I have completed the ClubProject form and made significant progress on the DSG Project form. Additionally, I refactored two controllers and multiple frontend components to improve their code quality and maintainability. This commit adds these changes to the codebase.

* chore:Bug fix

* chore:fix