feat: Implement landing page redirection on log out
Implement a feature that redirects users to the landing page when they log out of the system. Additionally, display a message to the user informing them that they have successfully logged out.

fix: Refresh club list when district is deleted

Fixed an issue where the club list was not being refreshed after a district was deleted. Now, when a district is deleted, the club list will be updated to reflect the changes.

feat: Enable site admin to add club members

Added a new feature that enables the site admin to add new members to a club. This will provide more flexibility and control to the site admin in managing club members.

chore: Associate Peter with Sunrise Club Cornwall

Added Peter as a member of the Sunrise Club Cornwall, as per the request of the club members. This will enable him to participate in club activities and events.